+++ camera/vdi_base/common/adapter/platform/v4l2/src/pipeline_core/nodes/uvc_node/uvc_node.cpp
@@ -102,9 +102,9 @@ RetCode UvcNode::Start(const int32_t streamId)
     for (const auto& it : outPorts) {
         DeviceFormat format;
         format.fmtdesc.pixelformat = V4L2_PIX_FMT_YUYV;
-        format.fmtdesc.width = wide_;
-        format.fmtdesc.height = high_;
-        int bufCnt = it->format_.bufferCount_;
+        format.fmtdesc.width = static_cast<uint32_t>(wide_);
+        format.fmtdesc.height = static_cast<uint32_t>(high_);
+        uint32_t bufCnt = it->format_.bufferCount_;
         rc = sensorController_->Start(bufCnt, format);
         if (rc == RC_ERROR) {
             CAMERA_LOGE("Start failed.");
@@ -164,7 +164,7 @@ void UvcNode::GetUpdateFps(const std::shared_ptr<CameraMetadata>& metadata)
     int ret = FindCameraMetadataItem(data, OHOS_CONTROL_FPS_RANGES, &entry);
     if (ret == 0) {
         std::vector<int32_t> fpsRange;
-        for (int i = 0; i < entry.count; i++) {
+        for (uint32_t i = 0; i < entry.count; i++) {
             fpsRange.push_back(*(entry.data.i32 + i));
         }
         meta_->addEntry(OHOS_CONTROL_FPS_RANGES, fpsRange.data(), fpsRange.size());
